Initial commit
Initial commit of project files.

UWP application which performs several actions for each TopCoder problem class added.
It opens the practice problem webpage in the TopCoder Arena.
It displays the code written to solve the problem.
It displays pass/fail and time elapsed for each example.
It displays pass/fail and time elapsed for each example over an average of several runs.