Added views and view models for several elements including the problem selection control, the progress bar and the text display area.
Incorporated the ProblemsList enum and the ProblemFactory functionality in to a single Problems class.
Re-added the method to return a problem from an index, now using the method to return the problem associated with that index in the ProblemsList.

Renamed ViewModel to MainPageVM and moved this to the new ViewModels folder.
Adapted MainPageVM to use the child viewmodels.
Adapted MainPage to use the new controls.